# log learning rate to DistriOptimizer and LocalOptimizer